backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1
google-genai==1.14.0
python-multipart==0.0.9
python-socketio[asgi]==5.11.0
httpx==0.28.1
google-generativeai==0.8.0
tavily-python==0.5.0
APScheduler==3.10.4